move admin API tests into test/api (#2086)
* move admin API tests into test/api

* move admin.py
