+++ tests/api/test_admin.py
@@ -39,14 +39,14 @@ from pathlib import Path
 import pytest
 
 from pygeoapi.util import yaml_load
-from pygeoapi.admin import (
+from pygeoapi.api.admin import (
     Admin, delete_resource, get_config_, get_resource,
     get_resources, patch_config, patch_resource, post_resource,
     put_config, put_resource)
 
 from tests.util import mock_api_request
 
-THISDIR = Path(__file__).resolve().parent
+THISDIR = Path(__file__).resolve().parent.parent
 
 
 @pytest.fixture()
